requirements.txt
flask
flask_restx
google-cloud-vision
boto3
pillow